requirements.txt
websockets==12.0
python-dotenv
pandas
numpy
requests